+++ nbproject/private/private.xml
@@ -3,9 +3,7 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/C:/Users/asmin/Desktop/TrabalhoFinal/Commit/TrabalhoFinal/Trabalho_Final/src/EntEmprestimo.java</file>
-            <file>file:/C:/Users/asmin/Desktop/TrabalhoFinal/Commit/TrabalhoFinal/Trabalho_Final/src/Principal.java</file>
-            <file>file:/C:/Users/asmin/Desktop/TrabalhoFinal/Commit/TrabalhoFinal/Trabalho_Final/src/ControlEmprestimo.java</file>
+            <file>file:/C:/Users/asmin/Desktop/TrabalhoFinal/Commit/COM220_trab_final/src/Biblioteca/Principal.java</file>
         </group>
     </open-files>
 </project-private>
