+++ src/Biblioteca/EntPublicacao.java
@@ -1,3 +1,5 @@
+package Biblioteca;
+
 
 import java.io.Serializable;
